pyproject.toml
[project]
name = "gitlab-populator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-gitlab>=6.3.0",
    "python-lorem>=1.3.0.post3",
    "tqdm>=4.67.1",
]
